requirements.txt
Flask
pinecone-client
